Fix integer truncation bug in superob BT calculation by refactoring to pure function (#90)
* parent 3c43d877d05915b8ad0500127cc1440fcca4d907
author Andy Stokely <amstokely> 1749142514 -0700
committer amstokely <andystokely@gmail.com> 1749842995 -0600

Changed name of datetime variable to dateTime and fixed bug that occurred when writing dateTime with superobs enabled in goes_abi.

Changed netcdf4 to netCDF4 in README instructions.

Fixed how Channel variable is set in goes_abi converter. Also added a function for creating fortran ctest tests.

Added units attribute for dateTime variable.

Updated test reference files.

Added docstring for  set_goes_abi_out_fname subroutine.

Fixed bug in read_HSD releated to the intermediate variable tb being an integer instead of a real.

Fixed bug in read_HSD releated to the intermediate variable tb being an integer instead of a real.

* Removed unused variable in hsd.f90

* Updated README

* Removed outdated build instructions.

* Updated HSD FLDK file conversion instructions.

* Updated HSD FLDK conversion instructions.

* Cleaned up C++ test names.

---------
