+++ test/cxx/CMakeLists.txt
@@ -4,9 +4,9 @@ FETCHCONTENT_DECLARE(
 )
 FETCHCONTENT_MAKEAVAILABLE(googletest)
 
-set(test_IodaObsSchema_SOURCES test_IodaObsSchema.cc fixture/IodaObsSchemaFixture.cc)
-list(TRANSFORM test_IodaObsSchema_SOURCES PREPEND ${CMAKE_CURRENT_SOURCE_DIR}/)
-set(test_IodaObsSchema_LIBRARIES yaml-cpp::yaml-cpp GTest::gtest_main obs2ioda_cxx)
-set(test_IodaObsSchema_INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_SOURCE_DIR}/obs2ioda-v3/src/cxx)
-add_cxx_ctest(test_IodaObsSchema "${test_IodaObsSchema_SOURCES}" "${test_IodaObsSchema_INCLUDE_DIRS}" "${test_IodaObsSchema_LIBRARIES}")
+set(test_ioda_obs_schema_SOURCES ioda_obs_schema.test.cc)
+list(TRANSFORM test_ioda_obs_schema_SOURCES PREPEND ${CMAKE_CURRENT_SOURCE_DIR}/)
+set(test_ioda_obs_schema_LIBRARIES yaml-cpp::yaml-cpp GTest::gtest_main obs2ioda_cxx)
+set(test_ioda_obs_schema_INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_SOURCE_DIR}/obs2ioda-v3/src/cxx)
+add_cxx_ctest(test_ioda_obs_schema "${test_ioda_obs_schema_SOURCES}" "${test_ioda_obs_schema_INCLUDE_DIRS}" "${test_ioda_obs_schema_LIBRARIES}")
